migrate commit callback tests
