+++ clients/src/test/java/org/apache/kafka/clients/consumer/internals/AsyncKafkaConsumerTest.java
@@ -16,20 +16,13 @@
  */
 package org.apache.kafka.clients.consumer.internals;
 
-import static java.util.Arrays.asList;
 import static java.util.Collections.singleton;
-import static java.util.Collections.singletonList;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.junit.jupiter.api.Assertions.fail;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.atLeast;
-import static org.mockito.Mockito.doAnswer;
-import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockConstruction;
 import static org.mockito.Mockito.never;
@@ -45,28 +38,16 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
-import java.util.stream.Stream;
 import org.apache.kafka.clients.ClientResponse;
-import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.clients.consumer.ConsumerRecords;
 import org.apache.kafka.clients.consumer.OffsetAndMetadata;
-import org.apache.kafka.clients.consumer.OffsetCommitCallback;
-import org.apache.kafka.clients.consumer.RetriableCommitFailedException;
 import org.apache.kafka.clients.consumer.internals.events.ApplicationEvent;
 import org.apache.kafka.clients.consumer.internals.events.ApplicationEventHandler;
 import org.apache.kafka.clients.consumer.internals.events.FetchCommittedOffsetsApplicationEvent;
 import org.apache.kafka.clients.consumer.internals.events.ResetPositionsApplicationEvent;
 import org.apache.kafka.clients.consumer.internals.events.ValidatePositionsApplicationEvent;
-import org.apache.kafka.common.KafkaException;
 import org.apache.kafka.common.Node;
 import org.apache.kafka.common.TopicPartition;
-import org.apache.kafka.common.errors.GroupAuthorizationException;
-import org.apache.kafka.common.errors.NetworkException;
-import org.apache.kafka.common.errors.RetriableException;
 import org.apache.kafka.common.errors.WakeupException;
 import org.apache.kafka.common.message.OffsetCommitResponseData;
 import org.apache.kafka.common.protocol.ApiKeys;
@@ -78,9 +59,6 @@ import org.apache.kafka.common.utils.Timer;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.function.Executable;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.ArgumentMatchers;
 import org.mockito.MockedConstruction;
 import org.mockito.stubbing.Answer;
@@ -88,10 +66,8 @@ import org.mockito.stubbing.Answer;
 public class AsyncKafkaConsumerTest {
 
     private AsyncKafkaConsumer<?, ?> consumer;
-    private FetchCollector<?, ?> fetchCollector;
     private ConsumerTestBuilder.AsyncKafkaConsumerTestBuilder testBuilder;
     private ApplicationEventHandler applicationEventHandler;
-    private SubscriptionState subscriptions;
 
     @BeforeEach
     public void setup() {
@@ -103,8 +79,6 @@ public class AsyncKafkaConsumerTest {
         testBuilder = new ConsumerTestBuilder.AsyncKafkaConsumerTestBuilder(groupInfo, enableAutoCommit, true);
         applicationEventHandler = testBuilder.applicationEventHandler;
         consumer = testBuilder.consumer;
-        fetchCollector = testBuilder.fetchCollector;
-        subscriptions = testBuilder.subscriptions;
     }
 
     @AfterEach
@@ -151,123 +125,6 @@ public class AsyncKafkaConsumerTest {
         assertEquals("This consumer has already been closed.", res.getMessage());
     }
 
-    @ParameterizedTest
-    @MethodSource("commitExceptionSupplier")
-    public void testCommitAsync_UserSuppliedCallback(Exception exception) {
-        CompletableFuture<Void> future = new CompletableFuture<>();
-
-        Map<TopicPartition, OffsetAndMetadata> offsets = new HashMap<>();
-        offsets.put(new TopicPartition("my-topic", 1), new OffsetAndMetadata(200L));
-
-        doReturn(future).when(consumer).commit(offsets, false);
-        MockCommitCallback callback = new MockCommitCallback();
-        assertDoesNotThrow(() -> consumer.commitAsync(offsets, callback));
-
-        if (exception == null) {
-            future.complete(null);
-            consumer.maybeInvokeCommitCallbacks();
-            assertNull(callback.exception);
-        } else {
-            future.completeExceptionally(exception);
-            consumer.maybeInvokeCommitCallbacks();
-            assertSame(exception.getClass(), callback.exception.getClass());
-        }
-    }
-
-    private static Stream<Exception> commitExceptionSupplier() {
-        return Stream.of(
-                null,  // For the successful completion scenario
-                new KafkaException("Test exception"),
-                new GroupAuthorizationException("Group authorization exception"));
-    }
-
-    @Test
-    public void testEnsureCallbackExecutedByApplicationThread() {
-        final String currentThread = Thread.currentThread().getName();
-        ExecutorService backgroundExecutor = Executors.newSingleThreadExecutor();
-        MockCommitCallback callback = new MockCommitCallback();
-        CountDownLatch latch = new CountDownLatch(1);  // Initialize the latch with a count of 1
-        try {
-            CompletableFuture<Void> future = new CompletableFuture<>();
-            doReturn(future).when(consumer).commit(new HashMap<>(), false);
-            assertDoesNotThrow(() -> consumer.commitAsync(new HashMap<>(), callback));
-            // Simulating some background work
-            backgroundExecutor.submit(() -> {
-                future.complete(null);
-                latch.countDown();
-            });
-            latch.await();
-            assertEquals(1, consumer.callbacks());
-            consumer.maybeInvokeCommitCallbacks();
-            assertEquals(currentThread, callback.completionThread);
-        } catch (Exception e) {
-            fail("Not expecting an exception");
-        } finally {
-            backgroundExecutor.shutdown();
-        }
-    }
-
-    @Test
-    public void testEnsureCommitSyncExecutedCommitAsyncCallbacks() {
-        MockCommitCallback callback = new MockCommitCallback();
-        CompletableFuture<Void> future = new CompletableFuture<>();
-        doReturn(future).when(consumer).commit(new HashMap<>(), false);
-        assertDoesNotThrow(() -> consumer.commitAsync(new HashMap<>(), callback));
-        future.completeExceptionally(new NetworkException("Test exception"));
-        assertMockCommitCallbackInvoked(() -> consumer.commitSync(),
-            callback,
-            Errors.NETWORK_EXCEPTION);
-    }
-
-
-    @Test
-    public void testEnsurePollExecutedCommitAsyncCallbacks() {
-        MockCommitCallback callback = new MockCommitCallback();
-        CompletableFuture<Void> future = new CompletableFuture<>();
-        consumer.assign(Collections.singleton(new TopicPartition("foo", 0)));
-        doReturn(future).when(consumer).commit(new HashMap<>(), false);
-        assertDoesNotThrow(() -> consumer.commitAsync(new HashMap<>(), callback));
-        future.complete(null);
-        assertMockCommitCallbackInvoked(() -> consumer.poll(Duration.ZERO),
-            callback,
-            null);
-    }
-
-    @Test
-    public void testEnsureShutdownExecutedCommitAsyncCallbacks() {
-        MockCommitCallback callback = new MockCommitCallback();
-        CompletableFuture<Void> future = new CompletableFuture<>();
-        doReturn(future).when(consumer).commit(new HashMap<>(), false);
-        assertDoesNotThrow(() -> consumer.commitAsync(new HashMap<>(), callback));
-        future.complete(null);
-        assertMockCommitCallbackInvoked(() -> consumer.close(),
-            callback,
-            null);
-    }
-
-    private void assertMockCommitCallbackInvoked(final Executable task,
-                                                 final MockCommitCallback callback,
-                                                 final Errors errors) {
-        assertDoesNotThrow(task);
-        assertEquals(1, callback.invoked);
-        if (errors == null)
-            assertNull(callback.exception);
-        else if (errors.exception() instanceof RetriableException)
-            assertTrue(callback.exception instanceof RetriableCommitFailedException);
-    }
-
-    private static class MockCommitCallback implements OffsetCommitCallback {
-        public int invoked = 0;
-        public Exception exception = null;
-        public String completionThread;
-
-        @Override
-        public void onComplete(Map<TopicPartition, OffsetAndMetadata> offsets, Exception exception) {
-            invoked++;
-            this.completionThread = Thread.currentThread().getName();
-            this.exception = exception;
-        }
-    }
     /**
      * This is a rather ugly bit of code. Not my choice :(
      *
